[Refactor] 모듈화 및 컴포넌트 분리 (디자인은 추후 재조정)
