+++ src/pages/my-test/components/ProblemView.module.scss
@@ -1,0 +1,15 @@
+.problem-content {
+  border: 1px solid #ccc;
+  border-radius: 8px;
+  padding: 20px;
+  background-color: #f8f9fa;
+  height: 100%;
+  overflow-y: auto;
+
+  pre {
+    white-space: pre-wrap;
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+    line-height: 1.6;
+    margin: 0;
+  }
+}
