+++ src/pages/my-test/components/TabSelector.module.scss
@@ -1,0 +1,24 @@
+.tabs {
+  display: flex;
+  gap: 10px;
+  margin-bottom: 15px;
+
+  .tab {
+    padding: 8px 16px;
+    border: none;
+    border-radius: 4px;
+    background-color: #f0f0f0;
+    cursor: pointer;
+    transition: background-color 0.2s, color 0.2s;
+    font-weight: 500;
+
+    &:hover {
+      background-color: #e0e0e0;
+    }
+
+    &.active {
+      background-color: #007bff;
+      color: white;
+    }
+  }
+}
