+++ src/pages/my-test/components/TeamCodeView.module.scss
@@ -1,0 +1,60 @@
+.team-code-content {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+
+  .member-navigation {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin-bottom: 15px;
+    padding: 8px;
+    background-color: #f8f9fa;
+    border-radius: 4px;
+
+    .member-indicator {
+      font-size: 14px;
+      font-weight: 500;
+    }
+
+    .nav-button {
+      padding: 5px 10px;
+      background-color: #007bff;
+      color: white;
+      border: none;
+      border-radius: 4px;
+      cursor: pointer;
+
+      &:hover {
+        background-color: #0069d9;
+      }
+
+      &.prev {
+        margin-right: auto;
+      }
+
+      &.next {
+        margin-left: auto;
+      }
+    }
+  }
+
+  .member-code-item {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+
+    .member-name {
+      font-size: 16px;
+      font-weight: 500;
+      margin-bottom: 10px;
+    }
+
+    .code-preview {
+      flex: 1;
+      border: 1px solid #e0e0e0;
+      border-radius: 4px;
+      overflow: hidden;
+    }
+  }
+}
