+++ src/pages/my-test/data/problemContent.ts
@@ -1,0 +1,18 @@
+// 삼각형 판별 문제의 내용
+export const problemContent = `
+    # 삼각형 판별하기
+
+    세 변의 길이가 주어졌을 때, 삼각형의 종류를 판별하는 프로그램을 작성하시오.
+
+    ## 입력
+    입력은 여러 개의 테스트 케이스로 이루어져 있다.
+    각 테스트 케이스는 세 정수 a, b, c로 이루어져 있으며, 각 변의 길이를 나타낸다.
+    입력의 마지막 줄에는 0 0 0이 주어진다.
+
+    ## 출력
+    각 테스트 케이스마다 다음과 같이 출력한다.
+    - 세 변의 길이가 삼각형의 조건을 만족하지 않으면 "Invalid"
+    - 세 변의 길이가 모두 같으면 "Equilateral"
+    - 두 변의 길이만 같으면 "Isosceles"
+    - 세 변의 길이가 모두 다르면 "Scalene"
+`;
